feat: backpack WIP: PR review: WearableCatalogGrid refactor adding WearableCatalogItem
